Renamed the module draw_events to visualiser.
